refactor signers
